Renaming InputParameters.{h|cpp} to Parameters.{h|cpp}
