+++ test/src/InitialCondition_test.cpp
@@ -2,7 +2,7 @@
 #include "LineMesh.h"
 #include "FENonlinearProblem_test.h"
 #include "InitialCondition.h"
-#include "InputParameters.h"
+#include "Parameters.h"
 #include "petsc.h"
 
 using namespace godzilla;
